First assignment
Enter 3 numbers from the keyboard
Find the product of the 3 numbers
Calculate the average of the 3 numbers. For this problem, just use
integer division and present the average as an integer.  We will deal
with real numbers later.
Use if statements and nested if statements to find the largest of the 3
numbers.
Provide appropriate screen prompts to describe the inputs and outputs.
Comment your Source code!